Supply an alternate implementation of ExerciseProviderTestScenarios compatible with <= go 1.6
