+++ providertests/providers_test.go
@@ -1,5 +1,3 @@
-// +build go1.7
-
 package providertests
 
 import (
