+++ providertests/utils.go
@@ -1,5 +1,3 @@
-// +build go1.7
-
 package providertests
 
 import (
@@ -7,11 +5,9 @@ import (
 	"encoding/base64"
 	"encoding/pem"
 	"io/ioutil"
-	"testing"
 
 	"github.com/russellhaering/gosaml2"
 	"github.com/russellhaering/goxmldsig"
-	"github.com/stretchr/testify/require"
 )
 
 func LoadXMLResponse(path string) string {
@@ -48,13 +44,4 @@ type ProviderTestScenario struct {
 	ScenarioName    string
 	Response        string
 	ServiceProvider *saml2.SAMLServiceProvider
-}
-
-func ExerciseProviderTestScenarios(t *testing.T, scenarios []ProviderTestScenario) {
-	for _, scenario := range scenarios {
-		t.Run(scenario.ScenarioName, func(t *testing.T) {
-			_, err := scenario.ServiceProvider.RetrieveAssertionInfo(scenario.Response)
-			require.NoError(t, err)
-		})
-	}
 }
